requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
langchain
python-dotenv
numpy
pydantic
guardrails-ai
